heatflow {microcalorimetry}	R Documentation

Fit microcalorimetry heat flow data and calculate microbal community growth parameters

Description

The function efficiently fits data with linear or non-linear model (generally with R2 > 0.99), returns P0/estimated.P0 (micrbial innitial biomass), k (microbial growth activity), dorment/active microbial biomass (parameter: nonactive/active), Pmax (maximum potential activety), and Std.Error of parameters.

Usage


heatflow=(input,nl_model=F,active_part_given=F,bootstrap=T,boot.N=1000,nrow.facet_wrap=NULL,
                  gaps=c("10s","60s","120s"),time.unit=c("s","min","h"),flow.unit=c("W","mW","uW"))

Arguments

input
heatflow data. can be matrix/dataframe, list(data1,data2,...), or a directory cntains multiple text files. the format of data matrix should bee same with data heat (row 1: active.part, number or NA; row 2 to 3: start and end baseline, if the data were already substracted by baseline supply 0.)

nl_model
logical. If TRUE, fit heat data with nonlinear model function; if FALSE, fit linear-transformed heat flow data with linear model function. Suggest the default linear fit

active_part_given
logical. If TRUE, fit heat data with active part known (usually use the base heat flow rate (P0) of control experiment as the known active part); if FALSE, both active and nonactive (dorment) part were estimated by fitting. Suggest the default linear fitting without known active part

bootstrap
logical. If TRUE, estimate Std.Error of parameters with bootstrap. Default TRUE

boot.N
Resample times for bootstrapping. Larger number returns better estimates but requires more run time. Default 100.

gaps
vector or character ('number' + 's'). "10s" suggests collecting data every 10 seconds. Higher density of data collection returns better fittings and estimates but consume more run time

time.unit
The time unit of origin heat data. Generally "s" (second). This fuction finally transform the time unit of data to 'h' (hour)

flow.unit
The unit of heat flow rate. Generally "uW" (10^-6 W). This fuction finally transform the unit of heat flow rate to 'uW'

nrow.facet_wrap
Number of columns of subfigers. see ggplot and facet_wrap in library ggplot2

Author(s)

Ruirui Chen and Zhongwang Jing

Examples


  library(ggplot2)

  #data mode: matrix
  data(heat)

  re=heatflow(heat,boot.N=10)
  re$all_result$`60s`$data_t1$fit.method
  re$RE
  re$nonlinear_plot$`120s`
  re$linear_plot$`120s`
  re$nonlinear_plot$`60s`
  re$linear_plot$`10s`

  re=heatflow(heat,nl_model=F,active_part_given=T,boot.N = 10,gaps = '40s')
  re$RE
  re$nonlinear_plot$`40s`
  re$linear_plot$`40s`



  #data mode: directory
  new_dir=paste(getwd(),'/heat.data',sep='')
  mkdir(new_dir)
  write.table(heat, paste(new_dir,'/heat1.txt',sep=''),sep = '\t',row.names = T)
  write.table(heat, paste(new_dir,'/heat2.txt',sep=''),sep = '\t',row.names = T)

  #active part known
  re=heatflow(new_dir,nl_model=F,active_part_given=T,boot.N = 10,gaps = '40s')
  re$RE
  re$nonlinear_plot$`40s`

  #active part not given
  re=heatflow(new_dir,nl_model=T,active_part_given=F,boot.N = 10,gaps = '40s')
  re$RE
  re$nonlinear_plot$`40s`



  #data mode: list
  data(heat.data)
  names(heat.data)
  re=heatflow(heat.data,boot.N=10,gaps = c('10s','120s'))
  re$RE
  re$nonlinear_plot$`10s`
  re$nonlinear_plot$`120s`

  re$linear_plot$`10s`
  re$linear_plot$`120s`


[Package microcalorimetry version 1.0 Index]